# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para o Tech Challenge 2 da FIAP - SOAT",
        "title": "Lanchonete API - Tech Challenge 2",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/pedidos": {
            "post": {
                "description": "Cria um pedido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedido"
                ],
                "summary": "Cria um pedido",
                "parameters": [
                    {
                        "description": "Pedido",
                        "name": "pedido",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Pedido"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pedidos/listartodos": {
            "post": {
                "description": "Lista todos os pedidos presentes no banco",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedido"
                ],
                "summary": "Lista todos os pedidos no banco",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Pedido"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pedidos/{ID}": {
            "get": {
                "description": "Busca um pedido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedido"
                ],
                "summary": "Busca um pedido",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Número do pedido",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Pedido"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pedidos/{nroPedido}/status/{status}": {
            "put": {
                "description": "Atualizar um pedido",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pedido"
                ],
                "summary": "Atualiza um pedido a partir de sua Identificação",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Número do pedido",
                        "name": "nroPedido",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Novo Status do pedido",
                        "name": "status",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Pedido"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/produto": {
            "post": {
                "description": "Cria um produto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produto"
                ],
                "summary": "Cria um produto",
                "parameters": [
                    {
                        "description": "Produto",
                        "name": "produto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Produto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/produto/editar": {
            "post": {
                "description": "Edita um produto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produto"
                ],
                "summary": "Edita um produto",
                "parameters": [
                    {
                        "description": "Produto",
                        "name": "cliente",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.Produto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/produto/{id}": {
            "get": {
                "description": "Busca um produto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produto"
                ],
                "summary": "Busca um produto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id do produto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.Produto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove um produto",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produto"
                ],
                "summary": "Remove um produto",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id do produto",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/produtos": {
            "get": {
                "description": "Lista todos os produtos cadastrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produto"
                ],
                "summary": "Lista todos os produtos no banco",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Produto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/produtos/{categoria}": {
            "get": {
                "description": "Lista todos os produtos por categoria",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "produto"
                ],
                "summary": "Lista os produtos por categoria",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Categoria de produtos",
                        "name": "categoria",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Produto"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entities.CatProduto": {
            "type": "string",
            "enum": [
                "Lanche",
                "Acompanhamento",
                "Bebida",
                "Sobremesa"
            ],
            "x-enum-varnames": [
                "Lanche",
                "Acompanhamento",
                "Bebida",
                "Sobremesa"
            ]
        },
        "entities.Cliente": {
            "type": "object",
            "properties": {
                "cpf": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
...
        "entities.Pedido": {
            "type": "object",
            "properties": {
                "cliente": {
                    "$ref": "#/definitions/entities.Cliente"
                },
                "identificacao": {
                    "type": "string"
                },
                "personalizacao": {
                    "type": "string"
                },
                "produtos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Produto"
                    }
                },
                "status": {
                    "$ref": "#/definitions/entities.StatusPedido"
                },
                "statusPagamento": {
                    "type": "string"
                },
                "timeStamp": {
                    "type": "string",
                    "format": "date-time"
                },
                "total": {
                    "type": "number"
                },
                "ultimaAtualizacao": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
        "entities.Produto": {
            "type": "object",
            "properties": {
                "categoria": {
                    "$ref": "#/definitions/entities.CatProduto"
                },
                "descricao": {
                    "type": "string"
                },
                "identificacao": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "preco": {
                    "type": "number"
                }
            }
        },
        "entities.StatusPedido": {
            "type": "string",
            "enum": [
                "Pendente",
                "Recebido",
                "Em preparação",
                "Pronto",
                "Finalizado"
            ],
            "x-enum-varnames": [
                "Pendente",
                "Recebido",
                "EmPreparacao",
                "Pronto",
                "Finalizado"
            ]
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}